Rework the entire lib.
Use the new TLV interface almost everywhere. Focus on getting the
Open Message through together with a bunch of parameters.

This seems to work now. Further development needed on making this
easier to use and add documentation.

I'm also not sure what we need to export and what not. I.e. is
Parameter important, or do we only want to use TLV *inside* a
parameter.